requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.1.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-core==0.3.6
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langgraph==0.2.28
langgraph-checkpoint==1.0.12
langgraph-checkpoint-sqlite==1.0.4
langsmith==0.1.129
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.50.2
orjson==3.10.7
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.3
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.13.1
